requirements.txt
accelerate==1.2.1
altgraph==0.17.4
anyio==4.7.0
Appium-Python-Client==4.4.0
attrs==24.3.0
av==14.0.1
certifi==2024.12.14
charset-normalizer==3.4.0
filelock==3.16.1
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
macholib==1.16.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.2
pillow==11.0.0
psutil==6.1.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PySocks==1.7.1
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
selenium==4.27.1
setuptools==75.6.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.3
tokenizers==0.21.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.47.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.6
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
wsproto==1.2.0